Add files via upload
This repository contains code files (.py) as well as music files(.mp3) of my group project named "Music Generation using Evolutionary Computing".
In this project, we used Python as the coding language as it is easy to use and has many inbuilt libraries. There are many projects available on the topic of music creation but ours is different from them by technology. We have used Evolutionary computing techniques and specifically genetic algorithms to compose huge numbers of music. This repository has three files. First is the Python file for the Genetic algorithm i.e. "genetic.py" and this file contains the main code of the genetic algorithm which we need to apply to the sequence of input notes. The second file is for playing the music i.e. "audio.py" and the third is for storing the musical tunes i.e. "music.mp3".

For this project, we have used MySQL as a backend. Anyone who wants to run this repository first needs to create a database named "btech". In that database, we need to create a table named "music". The fields of that music tables are melody1(char(50)), melody2(char(50)), composition (varchar(100)), rating (int). 

Once you have created this database and tables in it, you need to run the genetic.py file first which will create N numbers of compositions of two musical segments and store them in the database. After that run the audio.py file which will access each musical tune from the database and run that music using a music.mp3 file.